Clear main
